pr feedback refactoring
