+++ src/main/java/bio/overture/ego/repository/queryspecification/ApplicationPermissionSpecification.java
@@ -2,9 +2,7 @@ package bio.overture.ego.repository.queryspecification;
 
 import bio.overture.ego.model.entity.ApplicationPermission;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Component;
 
 @Slf4j
-@Component
 public class ApplicationPermissionSpecification
     extends AbstractPermissionSpecification<ApplicationPermission> {}
