+++ src/main/java/bio/overture/ego/repository/queryspecification/GroupPermissionSpecification.java
@@ -18,9 +18,7 @@ package bio.overture.ego.repository.queryspecification;
 
 import bio.overture.ego.model.entity.GroupPermission;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Component;
 
 @Slf4j
-@Component
 public class GroupPermissionSpecification
     extends AbstractPermissionSpecification<GroupPermission> {}
