+++ src/main/java/bio/overture/ego/service/ApplicationPermissionService.java
@@ -140,24 +140,6 @@ public class ApplicationPermissionService
             .map(Group::getPermissions)
             .flatMap(Collection::stream)
             .collect(toUnmodifiableSet());
-
-    return resolveFinalPermissions(appPermissions, groupPermissions).stream()
-        .map(
-            x -> {
-              val ownerType = x.getOwner().getClass().getSimpleName().toUpperCase();
-              val accessLevel = x.getAccessLevel();
-              val policy = x.getPolicy();
-              val owner = x.getOwner();
-              val id = x.getId();
-
-              return ResolvedPermissionResponse.builder()
-                  .accessLevel(accessLevel)
-                  .ownerType(ownerType)
-                  .policy(policy)
-                  .owner(owner)
-                  .id(id)
-                  .build();
-            })
-        .collect(toUnmodifiableList());
+    return getResolvedPermissionsResponse(appPermissions, groupPermissions);
   }
 }
